# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: LibsClaw OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating LibsClaw HTTP APIs to /api/v1.
    Dynamic LibsClaw configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill. The config scope also
    grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local LibsClaw server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/knowledge-bases:
    get:
      tags: [Knowledge Base]
      summary: List knowledge bases
      operationId: listKnowledgeBases
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: refresh_stats
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Knowledge Base]
      summary: Create a knowledge base
      operationId: createKnowledgeBase
      x-astrbot-scope: kb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseCreateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}:
    get:
      tags: [Knowledge Base]
      summary: Get a knowledge base
      operationId: getKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Knowledge Base]
      summary: Update a knowledge base
      operationId: updateKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Knowledge Base]
      summary: Delete a knowledge base
      operationId: deleteKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/stats:
    get:
      tags: [Knowledge Base]
      summary: Get knowledge base stats
      operationId: getKnowledgeBaseStats
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/wiki/tree:
    get:
      tags: [Knowledge Base]
      summary: Get the knowledge base Wiki page tree
      operationId: getKnowledgeWikiTree
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/wiki/pages:
    get:
      tags: [Knowledge Base]
      summary: Get a knowledge base Wiki page
      operationId: getKnowledgeWikiPage
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Knowledge Base]
      summary: Create a knowledge base Wiki page
      operationId: createKnowledgeWikiPage
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeWikiPageCreateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Knowledge Base]
      summary: Update a knowledge base Wiki page
      operationId: updateKnowledgeWikiPage
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeWikiPageUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Knowledge Base]
      summary: Delete a knowledge base Wiki page
      operationId: deleteKnowledgeWikiPage
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/wiki/rebuild:
    post:
      tags: [Knowledge Base]
      summary: Rebuild the knowledge base Wiki indexes
      operationId: rebuildKnowledgeWikiIndex
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/wiki/move:
    post:
      tags: [Knowledge Base]
      summary: Move a knowledge base Wiki page or directory
      operationId: moveKnowledgeWikiPath
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeWikiMoveRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/wiki/paths:
    delete:
      tags: [Knowledge Base]
      summary: Delete a knowledge base Wiki page or directory
      operationId: deleteKnowledgeWikiPath
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: recursive
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/wiki/import:
    post:
      tags: [Knowledge Base]
      summary: Import Markdown files, folders, or ZIP archives into the Wiki
      operationId: importKnowledgeWiki
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/KnowledgeWikiImportRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/wiki/graph:
    get:
      tags: [Knowledge Base]
      summary: Get the knowledge base Wiki graph
      operationId: getKnowledgeWikiGraph
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents:
    get:
      tags: [Knowledge Base]
      summary: List knowledge base documents
      operationId: listKnowledgeDocuments
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Filter documents by name (case-insensitive partial match).
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Knowledge Base]
      summary: Upload a document to a knowledge base
      operationId: uploadKnowledgeDocument
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/KnowledgeDocumentUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents/import:
    post:
      tags: [Knowledge Base]
      summary: Import documents already available on the server
      operationId: importKnowledgeDocuments
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeDocumentImportRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents/import-url:
    post:
      tags: [Knowledge Base]
      summary: Import a document from URL
      operationId: importKnowledgeDocumentFromUrl
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeDocumentUrlImportRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents/{document_id}:
    get:
      tags: [Knowledge Base]
      summary: Get a knowledge base document
      operationId: getKnowledgeDocument
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Knowledge Base]
      summary: Delete a knowledge base document
      operationId: deleteKnowledgeDocument
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/chunks:
    get:
      tags: [Knowledge Base]
      summary: List document chunks
      operationId: listKnowledgeChunks
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - name: document_id
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/chunks/{chunk_id}:
    delete:
      tags: [Knowledge Base]
      summary: Delete a document chunk
      operationId: deleteKnowledgeChunk
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/ChunkId"
        - name: document_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/retrieve:
    post:
      tags: [Knowledge Base]
      summary: Retrieve knowledge base chunks
      operationId: retrieveKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeRetrieveRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/tasks/{task_id}:
    get:
      tags: [Knowledge Base]
      summary: Get knowledge base import task progress
      operationId: getKnowledgeTask
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/stats:
    get:
      tags: [Stats]
      summary: Get runtime statistics
      operationId: getStats
      x-astrbot-scope: system
      parameters:
        - name: offset_sec
          in: query
          schema:
            type: integer
            default: 86400
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard LibsClaw success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
...
    KbId:
      name: kb_id
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  schemas:
...
    KnowledgeBaseRequest:
      type: object
      properties:
        kb_name:
          type: string
        description:
          type: string
        emoji:
          type: string
        embedding_provider_id:
          type: [string, "null"]
        rerank_provider_id:
          type: [string, "null"]
        chunk_size:
          type: integer
        chunk_overlap:
          type: integer
        top_k_dense:
          type: integer
        top_k_sparse:
          type: integer
        top_m_final:
          type: integer
      additionalProperties: false
    KnowledgeBaseCreateRequest:
      allOf:
        - $ref: "#/components/schemas/KnowledgeBaseRequest"
        - type: object
          required: [kb_name]
          properties:
            kb_name:
              type: string
            embedding_provider_id:
              type: [string, "null"]
    KnowledgeWikiPageCreateRequest:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
          description: Markdown page path relative to the knowledge directory.
        content:
          type: string
          description: Complete Markdown page content.
      additionalProperties: false
    KnowledgeWikiPageUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/KnowledgeWikiPageCreateRequest"
        - type: object
          required: [original_path]
          properties:
            original_path:
              type: string
              description: Existing page path when updating or renaming a page.
    KnowledgeWikiMoveRequest:
      type: object
      required: [source_path, target_path]
      properties:
        source_path:
          type: string
        target_path:
          type: string
      additionalProperties: false
    KnowledgeWikiImportRequest:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
          description: Markdown files or ZIP archives.
        paths:
          type: array
          items:
            type: string
          description: Wiki-relative path paired with each uploaded file.
        overwrite:
          type: boolean
          default: false
          description: Replace pages that already exist at the same path.
    KnowledgeDocumentUploadRequest:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
...